+++ CHANGELOG.md
@@ -24,6 +24,7 @@ changes you may need to make.
 - **BREAKING**: Array-style instances are no longer supported.
 - **BREAKING**: Remove deprecated CLI commands: `radarr`, `sonarr`, and `create-config`.
 - **BREAKING**: Removed `reset_unmatched_scores` support under quality profile score section.
+- **BREAKING**: Migration steps that dealt with the old `trash.yml` have been removed.
 
 ## [4.4.1] - 2023-04-08